Voice app functionality finally works - when any nav links including Recordings or Login are pressed again in the Recordings page, the microphone button duplicates -- need to look into that. Also, will need to add code for saving the recording
