+++ client/public/index.html
@@ -1,27 +1,26 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta name="description" content="Speech Therapy App"/>
 
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <head>
+
+        <meta charset="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <meta name="theme-color" content="#000000" />
+        <meta name="description" content="Speech Therapy App"/>
+        <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.9.0/css/all.css">
+        <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bootswatch/4.0.0/journal/bootstrap.min.css" />
+        <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.9.0/css/all.css">
 
+        <title>Speech Therapy</title>
 
-    <title>Speech Therapy</title>
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-
-
-<!-- MODAL FOR LOGIN -->
+    </head>
   
+    <body>
 
+        <noscript>You need to enable JavaScript to run this app.</noscript>
+        <div id="root"></div>
 
-
-
-  </body>
+    </body>
+        
 </html>
