+++ client/src/App.css
@@ -87,3 +87,7 @@ body {
     background: transparent;
 }
 
+
+.voiceContainer {
+    text-align: center;
+}
