+++ client/src/components/Pages/ViewSpeeches.js
@@ -5,6 +5,7 @@ import Speeches from "../SpeechCreation/voiceapp.js";
 function ViewSpeeches() {
     return (
       <div>
+          <p>Hallo</p>
             {/* ViewSpeeches */}
 
       </div>
